Solved the problem of too many relationships
Figured out why too many relationships were being created.
It turns out, for reasons I have yet to figure out, most likely going to post on the neo4j forms, that when I was calling my create relationship function, it ran on every node by default. Which is what I was looking for. I thought that I had to pass the guid every time to do this but neo4j did the matching all on its own, which is not only great but also confusing. So It seems like I only have to call the function. Now the question is do I have to keep passing procguid. In my mind it makes sense to pass it once to initate the process but I might not have to from the looks of it. I am going to deal with that later after running the other tests files I got.  Just to see how its running