Ran Test file
Made an interesting discovery. As I keep writing this I keep learning more about how the neo4j driver and more specifically neo4j works. So I do have to give it the process Id to fufuil but It will afect all nodes that match that query. So what does this mean. It means that all I have todo is somehow store the guids and then just iterate through them all to build the tree because from my test with the full log, relationships were made, just pertaining to that one guid. So I am getting expected action, now I just need to stretch it across all guids. That is a tomorrow problem to solve.